Io imports corrected. GRPC and BC Converter files added

+++ src/main/java/group19/ssd/p2p/Ledger.java
@@ -8,7 +8,7 @@ import java.util.Hashtable;
 
 public class Ledger {
     Hashtable<String, Long> users = new Hashtable<>();   //String vs Wallet
-    int minCoin = 15;
+    long minCoin = 15;
 
     public Ledger() {
         users.put(KademliaClient.publicKey, (long) minCoin);
